requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
requests==2.32.5
pytz==2025.2

# Optional but recommended for production
gunicorn==23.0.0
psycopg2-binary==2.9.12

# Development tools (optional)
black==24.10.0
flake8==7.1.1
pytest==8.3.4
pytest-django==4.8.1
